+++ kit/dapp/locales/ar-SA/tokens.json
@@ -249,7 +249,7 @@
         "noSchedules": "لم يتم تقديم جداول عائدات",
         "preparing": "جارٍ إعداد تكوين جدول العائدات...",
         "submitting": "جارٍ تقديم جدول العائدات...",
-        "success": "تم تعيين جدول العائدات بنجاح"
+        "success": "تم تكوين جدول العائدات بنجاح"
       },
       "submit": "تعيين الجدول",
       "submitting": "جارٍ تعيين الجدول...",
