+++ kit/dapp/locales/de-DE/tokens.json
@@ -249,7 +249,7 @@
         "noSchedules": "Es wurden keine Ertragszeitpläne bereitgestellt",
         "preparing": "Vorbereitung der Konfiguration des Ertragszeitplans...",
         "submitting": "Ertragszeitplan wird eingereicht...",
-        "success": "Ertragszeitplan erfolgreich festgelegt"
+        "success": "Ertragszeitplan erfolgreich konfiguriert"
       },
       "submit": "Zeitplan festlegen",
       "submitting": "Zeitplan wird festgelegt...",
