+++ kit/dapp/locales/en-US/tokens.json
@@ -268,9 +268,9 @@
         "failed": "Failed to set yield schedule",
         "invalidSchedule": "Invalid yield schedule configuration",
         "noSchedules": "No yield schedules provided",
-        "preparing": "Preparing yield schedule configuration...",
+        "preparing": "Preparing to configure yield schedule...",
         "submitting": "Submitting yield schedule...",
-        "success": "Yield schedule set successfully"
+        "success": "Yield schedule configured successfully"
       },
       "submit": "Set Schedule",
       "submitting": "Setting schedule...",
