+++ kit/dapp/locales/ja-JP/tokens.json
@@ -249,7 +249,7 @@
         "noSchedules": "利回りスケジュールが提供されていません",
         "preparing": "利回りスケジュール構成の準備中...",
         "submitting": "利回りスケジュールを送信中...",
-        "success": "利回りスケジュールが正常に設定されました"
+        "success": "利回りスケジュールが正常に構成されました"
       },
       "submit": "スケジュールを設定",
       "submitting": "スケジュールを設定中...",
